Added phpide:// links to line numbers
Line numbers in the left margin are now
hyperlinks to

    phpide://_PROJECTROOT_/file@linenum

With 'file' being the real path to the file
in the project, and linenum being the line number.

This will allow users to click on a line number and
have the corresponding file opened on the rigth line
in the IDE.
All they have to do is to configure their browser to
open phpide:// links with a specific script, that will
extract file + line, and run their IDE, passing it
those parameters as it expects.

Note : the project's root is different on each developer's
computer -- and, so, it's their local script that has to
replace '_PROJECTROOT_' by the right actual project root.

This should, more or less, fix issue #10