File lines must be displayed correctly whatever the file encoding is.
Symfony escaping is enabled and use htmlspecialchars internally (using sf_charset configuration directive to know the encoding)
The project is configure as UTF-8 project. Problems appear when lines of code to review are in another encoding (for example ISO-8859-1).
In that case, some lines can appear as empty.

The fix relies on encoding detection (at file level) to switch the sf_charset value (only to ISO-8859-1 for the moment).